+++ test/run-tests.js
@@ -1,6 +1,7 @@
 "use strict";
 
 require("./slmp-core.test");
+require("./slmp-capability-profiles.test");
 require("./slmp-high-level.test");
 require("./slmp-cpu-operation-state.test");
 require("./shared-vectors.test");